session: refactor transaction state.
In this PR TxnState is seperated from session package into a sub
package. And, many private methods of the struct is exported so that
session package can call them as previous.

The reason to do the refactor is I'm planning to support resolving
timestamp when execute instead of plan. The difference of these 2 phase
is when plan, TxnState can only be accessed by 1 thread, so that
synchronous is unnecessary. However when execute, TxnState can be
accessed by several task workers, so that the action from pending to
valid needs to be synchronized. This PR should make it more easily to
add the logic.
