+++ session/session.go
@@ -54,6 +54,7 @@ import (
 	"github.com/pingcap/tidb/plugin"
 	"github.com/pingcap/tidb/privilege"
 	"github.com/pingcap/tidb/privilege/privileges"
+	"github.com/pingcap/tidb/session/txnstate"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/sessionctx/binloginfo"
 	"github.com/pingcap/tidb/sessionctx/stmtctx"
@@ -155,7 +156,7 @@ func (h *StmtHistory) Count() int {
 type session struct {
 	// processInfo is used by ShowProcess(), and should be modified atomically.
 	processInfo atomic.Value
-	txn         TxnState
+	txn         txnstate.TxnState
 
 	mu struct {
 		sync.RWMutex
@@ -388,7 +389,7 @@ func (s *session) doCommit(ctx context.Context) error {
 		return nil
 	}
 	defer func() {
-		s.txn.changeToInvalid()
+		s.txn.ChangeToInvalid()
 		s.sessionVars.SetStatusFlag(mysql.ServerStatusInTrans, false)
 	}()
 	if s.txn.IsReadOnly() {
@@ -436,7 +437,7 @@ func (s *session) doCommit(ctx context.Context) error {
 func (s *session) doCommitWithRetry(ctx context.Context) error {
 	defer func() {
 		s.GetSessionVars().SetTxnIsolationLevelOneShotStateForNextTxn()
-		s.txn.changeToInvalid()
+		s.txn.ChangeToInvalid()
 		s.cleanRetryInfo()
 	}()
 	if !s.txn.Valid() {
@@ -536,7 +537,7 @@ func (s *session) RollbackTxn(ctx context.Context) {
 	if ctx.Value(inCloseSession{}) == nil {
 		s.cleanRetryInfo()
 	}
-	s.txn.changeToInvalid()
+	s.txn.ChangeToInvalid()
 	s.sessionVars.TxnCtx.Cleanup()
 	s.sessionVars.SetStatusFlag(mysql.ServerStatusInTrans, false)
 }
@@ -597,7 +598,7 @@ func (s *session) isTxnRetryableError(err error) bool {
 }
 
 func (s *session) checkTxnAborted(stmt sqlexec.Statement) error {
-	if s.txn.doNotCommit == nil {
+	if s.txn.DoNotCommit == nil {
 		return nil
 	}
 	// If the transaction is aborted, the following statements do not need to execute, except `commit` and `rollback`,
@@ -608,7 +609,7 @@ func (s *session) checkTxnAborted(stmt sqlexec.Statement) error {
 	if _, ok := stmt.(*executor.ExecStmt).StmtNode.(*ast.RollbackStmt); ok {
 		return nil
 	}
-	return errors.New("current transaction is aborted, commands ignored until end of transaction block:" + s.txn.doNotCommit.Error())
+	return errors.New("current transaction is aborted, commands ignored until end of transaction block:" + s.txn.DoNotCommit.Error())
 }
 
 func (s *session) retry(ctx context.Context, maxCnt uint) (err error) {
@@ -621,7 +622,7 @@ func (s *session) retry(ctx context.Context, maxCnt uint) (err error) {
 		if err != nil {
 			s.RollbackTxn(ctx)
 		}
-		s.txn.changeToInvalid()
+		s.txn.ChangeToInvalid()
 	}()
 
 	connID := s.sessionVars.ConnectionID
@@ -712,7 +713,7 @@ func (s *session) retry(ctx context.Context, maxCnt uint) (err error) {
 			zap.Error(err),
 			zap.String("txn", s.txn.GoString()))
 		kv.BackOff(retryCnt)
-		s.txn.changeToInvalid()
+		s.txn.ChangeToInvalid()
 		s.sessionVars.SetStatusFlag(mysql.ServerStatusInTrans, false)
 	}
 	return err
@@ -1230,7 +1231,7 @@ func (s *session) CachedPlanExec(ctx context.Context,
 	switch prepared.CachedPlan.(type) {
 	case *plannercore.PointGetPlan:
 		resultSet, err = stmt.PointGet(ctx, is)
-		s.txn.changeToInvalid()
+		s.txn.ChangeToInvalid()
 	case *plannercore.Update:
 		s.PrepareTSFuture(ctx)
 		s.GetSessionVars().StmtCtx.Priority = kv.PriorityHigh
@@ -1316,18 +1317,18 @@ func (s *session) DropPreparedStmt(stmtID uint32) error {
 }
 
 func (s *session) Txn(active bool) (kv.Transaction, error) {
-	if !s.txn.validOrPending() && active {
+	if !s.txn.ValidOrPending() && active {
 		return &s.txn, kv.ErrInvalidTxn
 	}
-	if s.txn.pending() && active {
+	if s.txn.Pending() && active {
 		// Transaction is lazy initialized.
 		// PrepareTxnCtx is called to get a tso future, makes s.txn a pending txn,
 		// If Txn() is called later, wait for the future to get a valid txn.
 		txnCap := s.getMembufCap()
-		if err := s.txn.changePendingToValid(txnCap); err != nil {
+		if err := s.txn.ChangePendingToValid(txnCap); err != nil {
 			logutil.BgLogger().Error("active transaction fail",
 				zap.Error(err))
-			s.txn.cleanup()
+			s.txn.Cleanup()
 			s.sessionVars.TxnCtx.StartTS = 0
 			return &s.txn, err
 		}
@@ -1405,7 +1406,7 @@ func (s *session) NewTxn(ctx context.Context) error {
 	if s.GetSessionVars().GetReplicaRead().IsFollowerRead() {
 		txn.SetOption(kv.ReplicaRead, kv.ReplicaReadFollower)
 	}
-	s.txn.changeInvalidToValid(txn)
+	s.txn.ChangeInvalidToValid(txn)
 	is := domain.GetDomain(s).InfoSchema()
 	s.sessionVars.TxnCtx = &variable.TransactionContext{
 		InfoSchema:    is,
@@ -1747,7 +1748,7 @@ func createSession(store kv.Storage) (*session, error) {
 	// session implements variable.GlobalVarAccessor. Bind it to ctx.
 	s.sessionVars.GlobalVarsAccessor = s
 	s.sessionVars.BinlogClient = binloginfo.GetPumpsClient()
-	s.txn.init()
+	s.txn.Default()
 	return s, nil
 }
 
@@ -1771,7 +1772,7 @@ func CreateSessionWithDomain(store kv.Storage, dom *domain.Domain) (*session, er
 	domain.BindDomain(s, dom)
 	// session implements variable.GlobalVarAccessor. Bind it to ctx.
 	s.sessionVars.GlobalVarsAccessor = s
-	s.txn.init()
+	s.txn.Default()
 	return s, nil
 }
 
@@ -1978,7 +1979,7 @@ func (s *session) loadCommonGlobalVariablesIfNeeded() error {
 // PrepareTxnCtx starts a goroutine to begin a transaction if needed, and creates a new transaction context.
 // It is called before we execute a sql query.
 func (s *session) PrepareTxnCtx(ctx context.Context) {
-	if s.txn.validOrPending() {
+	if s.txn.ValidOrPending() {
 		return
 	}
 
@@ -2000,12 +2001,12 @@ func (s *session) PrepareTxnCtx(ctx context.Context) {
 
 // PrepareTSFuture uses to try to get txn future.
 func (s *session) PrepareTSFuture(ctx context.Context) {
-	if !s.txn.validOrPending() {
+	if !s.txn.ValidOrPending() {
 		txnFuture := s.getTxnFuture(ctx)
-		s.txn.changeInvalidToPending(txnFuture)
-		s.GetSessionVars().TxnCtx.SetStmtFuture(txnFuture.future, 0)
+		s.txn.ChangeInvalidToPending(txnFuture)
+		s.GetSessionVars().TxnCtx.SetStmtFuture(txnFuture.GetFuture(), 0)
 	} else if s.GetSessionVars().IsPessimisticReadConsistency() {
-		s.GetSessionVars().TxnCtx.SetStmtFuture(s.getTxnFuture(ctx).future, 0)
+		s.GetSessionVars().TxnCtx.SetStmtFuture(s.getTxnFuture(ctx).GetFuture(), 0)
 	}
 }
 
@@ -2029,7 +2030,7 @@ func (s *session) InitTxnWithStartTS(startTS uint64) error {
 	if err != nil {
 		return err
 	}
-	s.txn.changeInvalidToValid(txn)
+	s.txn.ChangeInvalidToValid(txn)
 	s.txn.SetCap(s.getMembufCap())
 	err = s.loadCommonGlobalVariablesIfNeeded()
 	if err != nil {
