+++ session/tidb_test.go
@@ -172,30 +172,3 @@ func match(c *C, row []types.Datum, expected ...interface{}) {
 		c.Assert(got, Equals, need)
 	}
 }
-
-func (s *testMainSuite) TestKeysNeedLock(c *C) {
-	rowKey := tablecodec.EncodeRowKeyWithHandle(1, 1)
-	indexKey := tablecodec.EncodeIndexSeekKey(1, 1, []byte{1})
-	uniqueValue := make([]byte, 8)
-	uniqueUntouched := append(uniqueValue, '1')
-	nonUniqueVal := []byte{'0'}
-	nonUniqueUntouched := []byte{'1'}
-	var deleteVal []byte
-	rowVal := []byte{'a', 'b', 'c'}
-	tests := []struct {
-		key  []byte
-		val  []byte
-		need bool
-	}{
-		{rowKey, rowVal, true},
-		{rowKey, deleteVal, true},
-		{indexKey, nonUniqueVal, false},
-		{indexKey, nonUniqueUntouched, false},
-		{indexKey, uniqueValue, true},
-		{indexKey, uniqueUntouched, false},
-		{indexKey, deleteVal, false},
-	}
-	for _, tt := range tests {
-		c.Assert(keyNeedToLock(tt.key, tt.val), Equals, tt.need)
-	}
-}
